removedepartment
